backend/requirements.txt
fastapi
psycopg2-binary
uvicorn[standard]
python-dotenv
google-genai
gradio_client
rembg
python-multipart
celery[redis]
redis
sqlalchemy[asyncio]
asyncpg
motor
playwright
trimesh
numpy
opencv-python-headless
transformers
torch
Pillow
httpx
pymongo
